// packages/coding-agent/src/world/generated/resource_srpc.pb.ts
export const ResourceServiceDefinition = {
	typeName: "resource.ResourceService",
	methods: {
		/**
		 * ResourceClient starts an instance of a client for the ResourceService,
		 * yielding a new client ID. The client can use that ID for future RPCs
		 * accessing the Resource tree. When the streaming RPC ends, references to
		 * resources opened by the client will be released.
		 *
		 * @generated from rpc resource.ResourceService.ResourceClient
		 */
		ResourceClient: {
			name: "ResourceClient",
			I: ResourceClientRequest,
			O: ResourceClientResponse,
			kind: MethodKind.ServerStreaming,
		},
		/**
		 * ResourceRpc is a rpc request for an open resource handle.
		 * Exposes service(s) depending on the resource type.
		 * Component ID: resource_id from ResourceClient call.
		 *
		 * @generated from rpc resource.ResourceService.ResourceRpc
		 */
		ResourceRpc: {
			name: "ResourceRpc",
			I: RpcStreamPacket,
			O: RpcStreamPacket,
			kind: MethodKind.BiDiStreaming,
		},
		/**
		 * ResourceRefRelease releases a resource given a handle ID.
		 * This is called when no references remain to a resource.
		 *
		 * @generated from rpc resource.ResourceService.ResourceRefRelease
		 */
		ResourceRefRelease: {
			name: "ResourceRefRelease",
			I: ResourceRefReleaseRequest,
			O: ResourceRefReleaseResponse,
			kind: MethodKind.Unary,
		},
		/**
		 * ResourceRefAdopt acknowledges adoption of a resource created by a held RPC.
		 *
		 * @generated from rpc resource.ResourceService.ResourceRefAdopt
		 */
		ResourceRefAdopt: {
			name: "ResourceRefAdopt",
			I: ResourceRefAdoptRequest,
			O: ResourceRefAdoptResponse,
			kind: MethodKind.Unary,
		},
		/**
		 * ResourceAttach allows a client to provide resources that server-side
		 * RPC handlers can invoke via getAttachedRef(id). Session-only Init/Ack,
		 * then resources registered via Add/AddAck. After Init/Ack, mux_data
		 * carries yamux frames for all attached resources.
		 *
		 * @generated from rpc resource.ResourceService.ResourceAttach
		 */
		ResourceAttach: {
			name: "ResourceAttach",
			I: ResourceAttachRequest,
			O: ResourceAttachResponse,
			kind: MethodKind.BiDiStreaming,
		},
	},
} as const;

/**
 * ResourceService provides a set of stateful resources via an RPC service.
 *
 * @generated from service resource.ResourceService
 */
export interface ResourceService {
	/**
	 * ResourceClient starts an instance of a client for the ResourceService,
	 * yielding a new client ID. The client can use that ID for future RPCs
	 * accessing the Resource tree. When the streaming RPC ends, references to
	 * resources opened by the client will be released.
	 *
	 * @generated from rpc resource.ResourceService.ResourceClient
	 */
	ResourceClient(request: ResourceClientRequest, abortSignal?: AbortSignal): MessageStream<ResourceClientResponse>;

	/**
	 * ResourceRpc is a rpc request for an open resource handle.
	 * Exposes service(s) depending on the resource type.
	 * Component ID: resource_id from ResourceClient call.
	 *
	 * @generated from rpc resource.ResourceService.ResourceRpc
	 */
	ResourceRpc(request: MessageStream<RpcStreamPacket>, abortSignal?: AbortSignal): MessageStream<RpcStreamPacket>;

	/**
	 * ResourceRefRelease releases a resource given a handle ID.
	 * This is called when no references remain to a resource.
	 *
	 * @generated from rpc resource.ResourceService.ResourceRefRelease
	 */
	ResourceRefRelease(
		request: ResourceRefReleaseRequest,
		abortSignal?: AbortSignal,
	): Promise<ResourceRefReleaseResponse>;

	/**
	 * ResourceRefAdopt acknowledges adoption of a resource created by a held RPC.
	 *
	 * @generated from rpc resource.ResourceService.ResourceRefAdopt
	 */
	ResourceRefAdopt(request: ResourceRefAdoptRequest, abortSignal?: AbortSignal): Promise<ResourceRefAdoptResponse>;

	/**
	 * ResourceAttach allows a client to provide resources that server-side
	 * RPC handlers can invoke via getAttachedRef(id). Session-only Init/Ack,
	 * then resources registered via Add/AddAck. After Init/Ack, mux_data
	 * carries yamux frames for all attached resources.
	 *
	 * @generated from rpc resource.ResourceService.ResourceAttach
	 */
	ResourceAttach(
		request: MessageStream<ResourceAttachRequest>,
		abortSignal?: AbortSignal,
	): MessageStream<ResourceAttachResponse>;
}

export const ResourceServiceServiceName = ResourceServiceDefinition.typeName;

export class ResourceServiceClient implements ResourceService {
	private readonly rpc: ProtoRpc;
	private readonly service: string;
	constructor(rpc: ProtoRpc, opts?: { service?: string }) {
		this.service = opts?.service || ResourceServiceServiceName;
		this.rpc = rpc;
		this.ResourceClient = this.ResourceClient.bind(this);
		this.ResourceRpc = this.ResourceRpc.bind(this);
		this.ResourceRefRelease = this.ResourceRefRelease.bind(this);
		this.ResourceRefAdopt = this.ResourceRefAdopt.bind(this);
		this.ResourceAttach = this.ResourceAttach.bind(this);
	}
	/**
	 * ResourceClient starts an instance of a client for the ResourceService,
	 * yielding a new client ID. The client can use that ID for future RPCs
	 * accessing the Resource tree. When the streaming RPC ends, references to
	 * resources opened by the client will be released.
	 *
	 * @generated from rpc resource.ResourceService.ResourceClient
	 */
	ResourceClient(request: ResourceClientRequest, abortSignal?: AbortSignal): MessageStream<ResourceClientResponse> {
		const requestMsg = ResourceClientRequest.create(request);
		const result = this.rpc.serverStreamingRequest(
			this.service,
			ResourceServiceDefinition.methods.ResourceClient.name,
			ResourceClientRequest.toBinary(requestMsg),
			abortSignal || undefined,
		);
		return buildDecodeMessageTransform(ResourceClientResponse)(result);
	}

	/**
	 * ResourceRpc is a rpc request for an open resource handle.
	 * Exposes service(s) depending on the resource type.
	 * Component ID: resource_id from ResourceClient call.
	 *
	 * @generated from rpc resource.ResourceService.ResourceRpc
	 */
	ResourceRpc(request: MessageStream<RpcStreamPacket>, abortSignal?: AbortSignal): MessageStream<RpcStreamPacket> {
		const result = this.rpc.bidirectionalStreamingRequest(
			this.service,
			ResourceServiceDefinition.methods.ResourceRpc.name,
			buildEncodeMessageTransform(RpcStreamPacket)(request),
			abortSignal || undefined,
		);
		return buildDecodeMessageTransform(RpcStreamPacket)(result);
	}

	/**
	 * ResourceRefRelease releases a resource given a handle ID.
	 * This is called when no references remain to a resource.
	 *
	 * @generated from rpc resource.ResourceService.ResourceRefRelease
	 */
	async ResourceRefRelease(
		request: ResourceRefReleaseRequest,
		abortSignal?: AbortSignal,
	): Promise<ResourceRefReleaseResponse> {
		const requestMsg = ResourceRefReleaseRequest.create(request);
		const result = await this.rpc.request(
			this.service,
			ResourceServiceDefinition.methods.ResourceRefRelease.name,
			ResourceRefReleaseRequest.toBinary(requestMsg),
			abortSignal || undefined,
		);
		return ResourceRefReleaseResponse.fromBinary(result);
	}

	/**
	 * ResourceRefAdopt acknowledges adoption of a resource created by a held RPC.
	 *
	 * @generated from rpc resource.ResourceService.ResourceRefAdopt
	 */
	async ResourceRefAdopt(
		request: ResourceRefAdoptRequest,
		abortSignal?: AbortSignal,
	): Promise<ResourceRefAdoptResponse> {
		const requestMsg = ResourceRefAdoptRequest.create(request);
		const result = await this.rpc.request(
			this.service,
			ResourceServiceDefinition.methods.ResourceRefAdopt.name,
			ResourceRefAdoptRequest.toBinary(requestMsg),
			abortSignal || undefined,
		);
		return ResourceRefAdoptResponse.fromBinary(result);
	}

	/**
	 * ResourceAttach allows a client to provide resources that server-side
	 * RPC handlers can invoke via getAttachedRef(id). Session-only Init/Ack,
	 * then resources registered via Add/AddAck. After Init/Ack, mux_data
	 * carries yamux frames for all attached resources.
	 *
	 * @generated from rpc resource.ResourceService.ResourceAttach
	 */
	ResourceAttach(
		request: MessageStream<ResourceAttachRequest>,
		abortSignal?: AbortSignal,
	): MessageStream<ResourceAttachResponse> {
		const result = this.rpc.bidirectionalStreamingRequest(
			this.service,
			ResourceServiceDefinition.methods.ResourceAttach.name,
			buildEncodeMessageTransform(ResourceAttachRequest)(request),
			abortSignal || undefined,
		);
		return buildDecodeMessageTransform(ResourceAttachResponse)(result);
	}
}